0.0.3.287
modification to the Sonic Inventory handler, moved into a single list instead of the item as we realised this could cause issues down the line.
added function to save and load SoniceInventory list to a file in the minecraft world directory.
command added to give better access to that list, providing options to create, reset or remove entier sonic inventories
some changes to how SonicPackets are handled. this is incase the player requires an inventory which they do not have stored!